Moved configuration related files to core/configuration
