+++ src/core/configuration/config-file-schema.ts
